move experience context to context folder, change context provider
